ml_service/requirements.txt
fastapi
uvicorn
python-multipart
pillow
torch
torchvision
tensorflow
ultralytics